pyproject.toml
[project]
name = "codegen-cli"
version = "0.4.0"
description = "Command-line coding agent with iterative reasoning, inspired by Claude Code and Cursor CLI, powered by Google Gemini API."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "CodeGen CLI Team", email = "vaishnavipullakhandam@gmail.com"}
]
maintainers = [
    {name = "CodeGen CLI Team", email = "vaishnavipullakhandam@gmail.com"}
]
keywords = [
    "cli", "coding", "agent", "ai", "assistant", "code", "development", "programming",
    "python", "javascript", "go", "rust", "universal", "natural-language"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General",
    "Topic :: Utilities",
]
dependencies = [
    "beautifulsoup4>=4.13.5",
    "google-genai>=1.50.0,<2.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.0",
    "pydantic>=2.10.0",
    "instructor>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
]

[project.urls]
Homepage = "https://github.com/vaishnavip-23/CodeGen-CLI"
Documentation = "https://github.com/vaishnavip-23/CodeGen-CLI/blob/main/docs/README.md"
Repository = "https://github.com/vaishnavip-23/CodeGen-CLI"


[project.scripts]
codegen = "codegen_cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["codegen_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "codegen_cli/",
    "docs/",
    "LICENSE",
    "pyproject.toml",
]
